% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data_v2.0.R
\name{analyze_data_v2.0}
\alias{analyze_data_v2.0}
\title{analyze each segment of data by calling an internal function analyze_segment().}
\usage{
analyze_data_v2.0(data)
}
\description{
analyze each segment of data by calling an internal function analyze_segment().
}
