% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelData_v2.0.R
\name{channelData_v2.0}
\alias{channelData_v2.0}
\title{extracts relevant channel data from all segments data, and calculates one mean value per condition, as well as delta tracing.}
\usage{
channelData_v2.0(data, variable)
}
\description{
extracts relevant channel data from all segments data, and calculates one mean value per condition, as well as delta tracing.
}
