% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data_v2.0.R
\name{export_data_v2.0}
\alias{export_data_v2.0}
\title{export the data into one xlsx file per patient.
Two options to export data:
    1. write_xlsx() : it only works if JVM is installed.
   2. rio::export()}
\usage{
export_data_v2.0(data, myFile)
}
\description{
export the data into one xlsx file per patient.
Two options to export data:
    1. write_xlsx() : it only works if JVM is installed.
   2. rio::export()
}
