% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_v2.0.R
\name{interpolation_v2.0}
\alias{interpolation_v2.0}
\title{prepares data for linear interpolation by calling a function linear_interp_v2.0'.}
\usage{
interpolation_v2.0(data, rate, interpolation_time, trial_length)
}
\description{
prepares data for linear interpolation by calling a function linear_interp_v2.0'.
}
