% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_interp_v2.0.R
\name{linear_interp_v2.0}
\alias{linear_interp_v2.0}
\title{Interpolate all data columns to a specified length using linear interpolation.}
\usage{
linear_interp_v2.0(data, data2)
}
\description{
Interpolate all data columns to a specified length using linear interpolation.
}
