% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTracings3G_v2.0.R
\name{plotTracings3G_v2.0}
\alias{plotTracings3G_v2.0}
\title{makes bar plots along with tracing plots.
this assumes data of 3 groups with 2 conditions.}
\usage{
plotTracings3G_v2.0(
  groupAvg_channelTracings,
  channelTracings_groups,
  deltaChannel_groups,
  avgChannelvalues_groups,
  groupNames,
  trial_length,
  conditionNames,
  refGroup
)
}
\description{
makes bar plots along with tracing plots.
this assumes data of 3 groups with 2 conditions.
}
