% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_v2.0.R
\name{preprocess_v2.0}
\alias{preprocess_v2.0}
\title{tidy up data to pass to interpolation function.
it call function interpolation_v2.0 for interpolation purpose.}
\usage{
preprocess_v2.0(data, rate, interpolation_time, trial_length)
}
\description{
tidy up data to pass to interpolation function.
it call function interpolation_v2.0 for interpolation purpose.
}
