% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingAnalysisExport_v2.0.R
\name{processingAnalysisExport_v2.0}
\alias{processingAnalysisExport_v2.0}
\title{this performs pre-processing, analysis, and final exports data in the excel sheets.
calls preprocess_v2.0 for pre-processing.
calls analyze_data_v2.0 for analysis.
calls export_data_v2.0 to export data to xlsx files.}
\usage{
processingAnalysisExport_v2.0(
  data,
  rate,
  interpolation_time,
  trial_length,
  myFile
)
}
\description{
this performs pre-processing, analysis, and final exports data in the excel sheets.
calls preprocess_v2.0 for pre-processing.
calls analyze_data_v2.0 for analysis.
calls export_data_v2.0 to export data to xlsx files.
}
